FEBio2: Removed FEMaterialPoint::time and FEMaterialPoint::dt (won't be accessible in plugin anyways).
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBio2@8491 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBio2/Command.h
@@ -14,7 +14,7 @@ public:
 
 	void SetName(const char* szname)
 	{
-		int l = strlen(szname);
+		int l = (int)strlen(szname);
 		assert(l);
 		m_pszname = new char[l+1];
 		strcpy(m_pszname, szname);
@@ -22,7 +22,7 @@ public:
 
 	void SetDescription(const char* szdesc)
 	{
-		int l = strlen(szdesc);
+		int l = (int)strlen(szdesc);
 		assert(l);
 		m_pszdesc = new char[l+1];
 		strcpy(m_pszdesc, szdesc);
