+++ FEBioPlot/FEBioPlotFile.cpp
@@ -534,7 +534,7 @@ bool FEBioPlotFile::WriteMaterials(FEModel& fem)
 
 			// Make sure that the material name fits in the buffer
 			const char* sz = pm->GetName();
-			int l = strlen(sz);
+			int l = (int)strlen(sz);
 			if (l >= STR_SIZE) l = STR_SIZE - 1;
 			strncpy(szname, sz, l);
