+++ FECore/FEDomain2D.h
@@ -21,7 +21,7 @@ public:
     void Create(int nelems, int elemType);
     
     //! return nr of elements
-    int Elements() const { return m_Elem.size(); }
+    int Elements() const { return (int)m_Elem.size(); }
     
     //! element access
     FEElement2D& Element(int n) { return m_Elem[n]; }
